Part 3: Using NodeView for CompactRCF

+++ Java/core/src/main/java/com/amazon/randomcutforest/tree/INodeView.java
@@ -26,7 +26,7 @@ public interface INodeView<P> {
 
     boolean leafPointEquals(P point);
 
-    short getCutDimension();
+    int getCutDimension();
 
     double[] getLeafPoint();
